SpreadsheetMetadataLoaderTesting2 extracted SpreadsheetMetadataLoaderTesting
