+++ src/main/java/walkingkooka/spreadsheet/convert/SpreadsheetConverterContextTesting.java
@@ -17,7 +17,7 @@
 
 package walkingkooka.spreadsheet.convert;
 
-import walkingkooka.spreadsheet.meta.SpreadsheetMetadataLoaderTesting;
+import walkingkooka.spreadsheet.meta.SpreadsheetMetadataLoaderTesting2;
 import walkingkooka.spreadsheet.provider.SpreadsheetProviderContextTesting;
 import walkingkooka.spreadsheet.reference.SpreadsheetLabelNameResolverTesting;
 import walkingkooka.storage.convert.StorageConverterContextTesting;
@@ -25,7 +25,7 @@ import walkingkooka.tree.json.convert.JsonNodeConverterContextTesting;
 
 public interface SpreadsheetConverterContextTesting<C extends SpreadsheetConverterContext> extends JsonNodeConverterContextTesting<C>,
     SpreadsheetLabelNameResolverTesting<C>,
-    SpreadsheetMetadataLoaderTesting<C>,
+    SpreadsheetMetadataLoaderTesting2<C>,
     StorageConverterContextTesting<C>,
     SpreadsheetProviderContextTesting<C> {
